fix uint64 bug

+++ Source/Internal/Utility/AdClearIdUtil.swift
@@ -53,6 +53,6 @@ public struct AdClearId {
     private static func limitBitsTo(value: UInt64, maxBits: UInt64) -> UInt64 {
         let maxLength: UInt64 = (1 << maxBits) - 1
 
-        return value & maxLength
+        return value & UInt64(maxLength)
     }
 }
